+++ RadixWallet/Features/AppFeature/Overlay/SheetOverlayCoordinator+Reducer.swift
@@ -58,26 +58,3 @@ public struct SheetOverlayCoordinator: Sendable, FeatureReducer {
 		}
 	}
 }
-
-// MARK: - InfoLinkSheet
-public struct InfoLinkSheet: FeatureReducer {
-	public struct State: Sendable, Hashable, Identifiable {
-		public let id = UUID()
-		public let image: ImageAsset?
-		public let text: String
-	}
-
-	public enum ViewAction: Equatable, Sendable {
-		case infoLinkTapped(OverlayWindowClient.GlossaryItem)
-	}
-
-	@Dependency(\.overlayWindowClient) var overlayWindowClient
-
-	public func reduce(into state: inout State, viewAction: ViewAction) -> Effect<Action> {
-		switch viewAction {
-		case let .infoLinkTapped(infoLink):
-			overlayWindowClient.showInfoLink(infoLink)
-			return .none
-		}
-	}
-}
